Move the rest of core EAR classes into core package
