+++ ear-compiler/src/main/java/eflang/ear/composer/AbstractScaleComposer.java
@@ -1,6 +1,6 @@
 package eflang.ear.composer;
 
-import eflang.ear.Scale;
+import eflang.ear.core.Scale;
 
 public abstract class AbstractScaleComposer implements Composer {
     protected Scale scale;
