+++ ear-compiler/src/main/java/eflang/ear/composer/OnlyRunsComposer.java
@@ -1,6 +1,6 @@
 package eflang.ear.composer;
 
-import eflang.ear.Scale;
+import eflang.ear.core.Scale;
 
 import java.util.List;
