+++ ear-compiler/src/main/java/eflang/ear/core/Argument.java
@@ -1,4 +1,4 @@
-package eflang.ear;
+package eflang.ear.core;
 
 public class Argument {
     public enum Type {
