+++ ear-compiler/src/main/java/eflang/ear/core/Scales.java
@@ -1,4 +1,4 @@
-package eflang.ear;
+package eflang.ear.core;
 
 public class Scales {
