+++ ear-compiler/src/main/java/eflang/ear/operation/Add.java
@@ -1,6 +1,6 @@
 package eflang.ear.operation;
 
-import eflang.ear.Argument;
+import eflang.ear.core.Argument;
 import eflang.ear.core.Instruction;
 
 import java.util.ArrayList;
