+++ hammer/src/main/java/eflang/hammer/EarCodeSupplier.java
@@ -1,7 +1,7 @@
 package eflang.hammer;
 
 import eflang.ear.EARCompiler;
-import eflang.ear.Scales;
+import eflang.ear.core.Scales;
 import eflang.ear.composer.Composer;
 import eflang.ear.composer.OnlyRunsComposer;
