+++ vibe/src/main/java/eflang/vibe/VibeController.java
@@ -4,8 +4,8 @@ import eflang.core.Parser;
 import eflang.core.StringMusicSource;
 import eflang.ear.EARCompiler;
 import eflang.ear.EARException;
-import eflang.ear.Scale;
-import eflang.ear.Scales;
+import eflang.ear.core.Scale;
+import eflang.ear.core.Scales;
 import eflang.ear.composer.GeometricComposer;
 import eflang.lobe.LOBECompiler;
 import eflang.lobe.LobeCompilationException;
